+++ libs/designer/src/lib/core/actions/bjsworkflow/templates.ts
@@ -20,6 +20,7 @@ import {
   TemplateService,
   type Template,
   clone,
+  getTriggerFromDefinition,
 } from '@microsoft/logic-apps-shared';
 import { createAsyncThunk } from '@reduxjs/toolkit';
 import type { RootState } from '../../state/templates/store';
@@ -46,10 +47,12 @@ export interface WorkflowTemplateData {
     light?: string;
     dark?: string;
   };
+  triggerType: string;
   connectionKeys: string[];
   errors: {
     workflow: string | undefined;
     kind?: string;
+    triggerDescription?: string;
   };
 }
 
@@ -226,12 +229,22 @@ export const loadTemplate = createAsyncThunk(
 
 export const validateWorkflowsBasicInfo = createAsyncThunk(
   'validateWorkflowsBasicInfo',
-  async ({ existingWorkflowNames }: { existingWorkflowNames: string[] }, thunkAPI) => {
+  async (
+    { existingWorkflowNames, requireDescription = false }: { existingWorkflowNames: string[]; requireDescription?: boolean },
+    thunkAPI
+  ) => {
     const state: RootState = thunkAPI.getState() as RootState;
     const { subscriptionId, resourceGroup: resourceGroupName, isConsumption } = state.workflow;
     const { workflows } = state.template;
     const workflowIds = Object.keys(workflows);
-    const result: Record<string, { kindError?: string; nameError?: string }> = {};
+    const result: Record<
+      string,
+      {
+        kindError?: string;
+        nameError?: string;
+        triggerDescriptionError?: string;
+      }
+    > = {};
     if (workflowIds.length) {
       const intl = getIntl();
       for (const id of workflowIds) {
@@ -259,6 +272,18 @@ export const validateWorkflowsBasicInfo = createAsyncThunk(
           ...result[id],
           nameError,
         };
+
+        if (requireDescription) {
+          const triggerKey = Object.keys(workflows?.[id]?.workflowDefinition?.triggers ?? {})?.[0];
+          const trigger = workflows?.[id]?.workflowDefinition?.triggers?.[triggerKey];
+          const triggerDescriptionError = await validateTriggerDescription(trigger?.description);
+          if (triggerDescriptionError) {
+            result[id] = {
+              ...result[id],
+              triggerDescriptionError,
+            };
+          }
+        }
       }
     }
 
@@ -313,6 +338,18 @@ export const validateWorkflowName = async (
     return availabilityError;
   }
 
+  return undefined;
+};
+
+export const validateTriggerDescription = async (triggerDescription: string | undefined) => {
+  const intl = getIntl();
+  if (!triggerDescription) {
+    return intl.formatMessage({
+      defaultMessage: 'Must provide value for description.',
+      id: 'OZ42O1',
+      description: 'Error message when the description is empty.',
+    });
+  }
   return undefined;
 };
 
@@ -415,6 +452,7 @@ const loadWorkflowTemplate = async (
             : workflowManifest.kinds?.length
               ? workflowManifest.kinds[0]
               : 'stateful',
+        triggerType: getTriggerFromDefinition(templateWorkflowDefinition.triggers ?? {}),
         images: {
           light: TemplateService().getContentPathUrl(`${templateId}/${workflowId}`, workflowManifest.images.light),
           dark: TemplateService().getContentPathUrl(`${templateId}/${workflowId}`, workflowManifest.images.dark),