+++ libs/designer/src/lib/ui/configuretemplate/tabs/parametersTab.tsx
@@ -1,33 +1,26 @@
 import type { AppDispatch } from '../../../core/state/templates/store';
 import type { TemplateTabProps } from '@microsoft/designer-ui';
 import constants from '../../../common/constants';
-import type { IntlShape } from 'react-intl';
 import { selectWizardTab } from '../../../core/state/templates/tabSlice';
 import { TemplateParametersList } from '../parameters/parameterslist';
+import type { TemplateWizardTabProps } from './model';
 
-export const parametersTab = (intl: IntlShape, dispatch: AppDispatch): TemplateTabProps => ({
+export const parametersTab = (
+  resources: Record<string, string>,
+  dispatch: AppDispatch,
+  { disabled, tabStatusIcon }: TemplateWizardTabProps
+): TemplateTabProps => ({
   id: constants.CONFIGURE_TEMPLATE_WIZARD_TAB_NAMES.PARAMETERS,
-  title: intl.formatMessage({
-    defaultMessage: 'Parameters',
-    id: 'lYAlE9',
-    description: 'The tab label for the monitoring parameters tab on the configure template wizard',
-  }),
-  hasError: false,
+  title: resources.ParametersTabLabel,
+  tabStatusIcon,
+  disabled,
   content: <TemplateParametersList />,
   footerContent: {
-    primaryButtonText: intl.formatMessage({
-      defaultMessage: 'Previous',
-      id: 'Q1LEiE',
-      description: 'Button text for going back to the previous tab',
-    }),
+    primaryButtonText: resources.PreviousButtonText,
     primaryButtonOnClick: () => {
       dispatch(selectWizardTab(constants.CONFIGURE_TEMPLATE_WIZARD_TAB_NAMES.CONNECTIONS));
     },
-    secondaryButtonText: intl.formatMessage({
-      defaultMessage: 'Next',
-      id: 'daThty',
-      description: 'Button text for proceeding to the next tab',
-    }),
+    secondaryButtonText: resources.NextButtonText,
     secondaryButtonOnClick: () => {
       dispatch(selectWizardTab(constants.CONFIGURE_TEMPLATE_WIZARD_TAB_NAMES.PROFILE));
     },